requirements.txt
tensorflow>=2.10
numpy
pillow
matplotlib
streamlit
scipy
pandas
